requirements.txt
transformers
kaggle
sentencepiece
pandas
numpy
matplotlib
seaborn
pytorch-lightning
torchmetrics
torch
easydict
hydra-core